Extract config change test into its own test class
