+++ applications/workers/workers-smoketest/src/smokeTest/kotlin/net/corda/applications/workers/smoketest/flow/FlowTests.kt
@@ -15,34 +15,21 @@ import net.corda.e2etest.utilities.TEST_NOTARY_CPI_NAME
 import net.corda.e2etest.utilities.awaitRestFlowResult
 import net.corda.e2etest.utilities.conditionallyUploadCordaPackage
 import net.corda.e2etest.utilities.conditionallyUploadCpiSigningCertificate
-import net.corda.e2etest.utilities.config.configWithDefaultsNode
-import net.corda.e2etest.utilities.config.getConfig
-import net.corda.e2etest.utilities.config.managedConfig
-import net.corda.e2etest.utilities.config.waitForConfigurationChange
 import net.corda.e2etest.utilities.getHoldingIdShortHash
 import net.corda.e2etest.utilities.getOrCreateVirtualNodeFor
 import net.corda.e2etest.utilities.registerStaticMember
 import net.corda.e2etest.utilities.startRpcFlow
-import net.corda.schema.configuration.ConfigKeys.MESSAGING_CONFIG
-import net.corda.schema.configuration.MessagingConfig.MAX_ALLOWED_MSG_SIZE
 import net.corda.v5.crypto.DigestAlgorithmName
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Assertions.assertAll
 import org.junit.jupiter.api.BeforeAll
-import org.junit.jupiter.api.MethodOrderer
-import org.junit.jupiter.api.Order
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.TestInstance
 import org.junit.jupiter.api.TestInstance.Lifecycle
-import org.junit.jupiter.api.TestMethodOrder
 import java.util.UUID
 import kotlin.text.Typography.quote
 
 @Suppress("Unused", "FunctionName")
-//The flow tests must go last as one test updates the messaging config which is highly disruptive to subsequent test runs. The real
-// solution to this is a larger effort to have components listen to their messaging pattern lifecycle status and for them to go DOWN when
-// their patterns are DOWN - CORE-8015
-@Order(999)
 @TestInstance(Lifecycle.PER_CLASS)
 class FlowTests {
 
@@ -847,57 +834,6 @@ class FlowTests {
                 assertThat(consumeResult.flowError?.message).contains("Unable to notarize transaction")
                 assertThat(consumeResult.flowError?.message).contains("Reference State Conflict")
             })
-        }
-    }
-
-    @Test
-    fun `cluster configuration changes are picked up and workers continue to operate normally`() {
-        val currentConfigValue = getConfig(MESSAGING_CONFIG).configWithDefaultsNode()[MAX_ALLOWED_MSG_SIZE].asInt()
-        val newConfigurationValue = (currentConfigValue * 1.5).toInt()
-
-        managedConfig { configManager ->
-            println("Set new config")
-            configManager
-                .load(MESSAGING_CONFIG, MAX_ALLOWED_MSG_SIZE, newConfigurationValue)
-                .apply()
-            // Wait for the rpc-worker to reload the configuration and come back up
-            println("Wait for the rpc-worker to reload the configuration and come back up")
-            waitForConfigurationChange(MESSAGING_CONFIG, MAX_ALLOWED_MSG_SIZE, newConfigurationValue.toString(), false)
-
-            // Execute some flows which require functionality from different workers and make sure they succeed
-            println("Execute some flows which require functionality from different workers and make sure they succeed")
-            val flowIds = mutableListOf(
-                startRpcFlow(
-                    bobHoldingId,
-                    RpcSmokeTestInput().apply {
-                        command = "persistence_persist"
-                        data = mapOf("id" to UUID.randomUUID().toString())
-                    }
-                ),
-
-                startRpcFlow(
-                    bobHoldingId,
-                    RpcSmokeTestInput().apply {
-                        command = "crypto_sign_and_verify"
-                        data = mapOf("memberX500" to bobX500)
-                    }
-                ),
-
-                startRpcFlow(
-                    bobHoldingId,
-                    RpcSmokeTestInput().apply {
-                        command = "lookup_member_by_x500_name"
-                        data = mapOf("id" to charlyX500)
-                    }
-                )
-            )
-
-            println("Check status of flows")
-            flowIds.forEach {
-                val flowResult = awaitRestFlowResult(bobHoldingId, it)
-                assertThat(flowResult.flowError).isNull()
-                assertThat(flowResult.flowStatus).isEqualTo(RPC_FLOW_STATUS_SUCCESS)
-            }
         }
     }
